requirements.txt
annotated-types==0.7.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
clickhouse-connect==0.5.23
confluent-kafka==2.5.3
idna==3.10
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kafka-python==2.0.2
orjson==3.10.7
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
python-dotenv==1.0.1
quixstreams==2.11.0
referencing==0.35.1
requests==2.32.3
rocksdict==0.3.23
rpds-py==0.20.0
typing_extensions==4.12.2
urllib3==2.2.3